+++ client/js/verification-portal/community-health-center/validate.js
@@ -1,0 +1,54 @@
+import validator from 'email-validator'
+
+const validateEmail = email => {
+	return validator.validate(email)
+}
+
+const validate = ({
+	chcName,
+	addressLine1,
+	city,
+	state,
+	zipCode,
+	phoneNumber,
+	email,
+	website,
+}) => {
+	const errors = {}
+
+	if (!chcName) {
+		errors.chcName = 'Required'
+	}
+
+	if (!addressLine1) {
+		errors.addressLine1 = 'Required'
+	}
+
+	if (!city) {
+		errors.city = 'Required'
+	}
+
+	if (!state) {
+		errors.state = 'Required'
+	}
+
+	if (!zipCode) {
+		errors.zipCode = 'Required'
+	}
+
+	if (!phoneNumber) {
+		errors.phoneNumber = 'Required'
+	}
+
+	if (!website) {
+		errors.website = 'Required'
+	}
+
+	if (email && !validateEmail(email)) {
+		errors.email = 'Please enter a valid email address'
+	}
+
+	return errors
+}
+
+export default validate
